fix all edsge case that could make pbdt chain stall
