+++ libraries/cli/include/cli/config_jsons/default/default_genesis.json
@@ -118,8 +118,6 @@
     "ficus_hf": {
       "block_num": 50,
       "pillar_blocks_interval": 16,
-      "pillar_chain_sync_interval": 8,
-      "pbft_inclusion_delay": 6,
       "bridge_contract_address": "0xcAF2b453FE8382a4B8110356DF0508f6d71F22BF"
     }
   }
