+++ libraries/cli/include/cli/config_jsons/devnet/devnet_genesis.json
@@ -283,8 +283,6 @@
     "ficus_hf": {
       "block_num": 1000,
       "pillar_blocks_interval": 100,
-      "pillar_chain_sync_interval": 25,
-      "pbft_inclusion_delay": 6,
       "bridge_contract_address": "0xcAF2b453FE8382a4B8110356DF0508f6d71F22BF"
     }
   }
