+++ libraries/cli/include/cli/config_jsons/mainnet/mainnet_genesis.json
@@ -1652,8 +1652,6 @@
     "ficus_hf": {
       "block_num": -1,
       "pillar_blocks_interval": 100,
-      "pillar_chain_sync_interval": 25,
-      "pbft_inclusion_delay": 6,
       "bridge_contract_address": "0x0000000000000000000000000000000000000000"
     }
   }
