+++ libraries/config/src/hardfork.cpp
@@ -59,7 +59,8 @@ bool FicusHardforkConfig::isPillarBlockPeriod(taraxa::PbftPeriod period, bool sk
 }
 
 bool FicusHardforkConfig::isPbftWithPillarBlockPeriod(taraxa::PbftPeriod period) const {
-  return period >= firstPillarBlockPeriod() && period % pillar_blocks_interval == pbft_inclusion_delay;
+  // Pillar block hash is included in the next pbft block with period +1
+  return period >= firstPillarBlockPeriod() && period % pillar_blocks_interval == 1;
 }
 
 // Returns first pillar block period
@@ -73,20 +74,12 @@ void FicusHardforkConfig::validate(uint32_t delegation_delay) const {
     return;
   }
 
+  if (pillar_blocks_interval <= 1) {
+    throw taraxa::ConfigException("ficus_hf.pillar_blocks_interval must be > 1");
+  }
+
   if (block_num % pillar_blocks_interval) {
     throw taraxa::ConfigException("ficus_hf.block_num % ficus_hf.pillar_blocks_interval must == 0");
-  }
-
-  if (pillar_blocks_interval <= pillar_chain_sync_interval) {
-    throw taraxa::ConfigException("ficus_hf.pillar_blocks_interval must be > ficus_hf.pillar_chain_sync_interval");
-  }
-
-  if (pillar_chain_sync_interval <= pbft_inclusion_delay) {
-    throw taraxa::ConfigException("ficus_hf.pillar_chain_sync_interval must be > ficus_hf.pbft_inclusion_delay");
-  }
-
-  if (pbft_inclusion_delay < 1 || pbft_inclusion_delay <= delegation_delay) {
-    throw taraxa::ConfigException("ficus_hf.pbft_inclusion_delay must be >= 1 && > dpos.delegation_delay");
   }
 }
 
@@ -94,8 +87,6 @@ Json::Value enc_json(const FicusHardforkConfig& obj) {
   Json::Value json(Json::objectValue);
   json["block_num"] = dev::toJS(obj.block_num);
   json["pillar_blocks_interval"] = dev::toJS(obj.pillar_blocks_interval);
-  json["pillar_chain_sync_interval"] = dev::toJS(obj.pillar_chain_sync_interval);
-  json["pbft_inclusion_delay"] = dev::toJS(obj.pbft_inclusion_delay);
   json["bridge_contract_address"] = dev::toJS(obj.bridge_contract_address);
   return json;
 }
@@ -103,13 +94,10 @@ Json::Value enc_json(const FicusHardforkConfig& obj) {
 void dec_json(const Json::Value& json, FicusHardforkConfig& obj) {
   obj.block_num = json["block_num"].isUInt64() ? dev::getUInt(json["block_num"]) : uint64_t(-1);
   obj.pillar_blocks_interval = dev::getUInt(json["pillar_blocks_interval"]);
-  obj.pillar_chain_sync_interval = dev::getUInt(json["pillar_chain_sync_interval"]);
-  obj.pbft_inclusion_delay = dev::getUInt(json["pbft_inclusion_delay"]);
   obj.bridge_contract_address = taraxa::addr_t(json["bridge_contract_address"].asString());
 }
 
-RLP_FIELDS_DEFINE(FicusHardforkConfig, block_num, pillar_blocks_interval, pillar_chain_sync_interval,
-                  pbft_inclusion_delay, bridge_contract_address)
+RLP_FIELDS_DEFINE(FicusHardforkConfig, block_num, pillar_blocks_interval, bridge_contract_address)
 
 // Json::Value enc_json(const BambooRedelegation& obj) {
 //   Json::Value json(Json::objectValue);
