+++ libraries/core_libs/consensus/include/pbft/pbft_manager.hpp
@@ -598,6 +598,7 @@ class PbftManager {
   bool already_next_voted_null_block_hash_ = false;
   bool go_finish_state_ = false;
   bool loop_back_finish_state_ = false;
+  bool already_placed_pillar_vote_ = false;
 
   // Used to avoid cyclic logging in voting steps that are called repeatedly
   bool printSecondFinishStepInfo_ = true;
