+++ libraries/core_libs/consensus/src/pbft/pbft_manager.cpp
@@ -17,7 +17,7 @@
 #include "vote_manager/vote_manager.hpp"
 
 namespace taraxa {
-using vrf_output_t = vrf_wrapper::vrf_output_t;
+using namespace std::chrono_literals;
 
 constexpr std::chrono::milliseconds kPollingIntervalMs{100};
 constexpr PbftStep kMaxSteps{13};  // Need to be a odd number
@@ -307,6 +307,7 @@ bool PbftManager::advancePeriod() {
   resetPbftConsensus(1 /* round */);
   broadcast_reward_votes_counter_ = 1;
   rebroadcast_reward_votes_counter_ = 1;
+  already_placed_pillar_vote_ = false;
   current_period_start_datetime_ = std::chrono::system_clock::now();
 
   const auto new_period = getPbftPeriod();
@@ -721,9 +722,10 @@ std::shared_ptr<PbftBlock> PbftManager::getValidPbftProposedBlock(PbftPeriod per
 
 bool PbftManager::placeVote(const std::shared_ptr<PbftVote> &vote, std::string_view log_vote_id,
                             const std::shared_ptr<PbftBlock> &voted_block) {
+  const PbftPeriod period = vote->getPeriod();
   if (!vote_mgr_->addVerifiedVote(vote)) {
     LOG(log_er_) << "Unable to place vote " << vote->getHash() << " for block " << vote->getBlockHash() << ", period "
-                 << vote->getPeriod() << ", round " << vote->getRound() << ", step " << vote->getStep();
+                 << period << ", round " << vote->getRound() << ", step " << vote->getStep();
     return false;
   }
 
@@ -733,8 +735,8 @@ bool PbftManager::placeVote(const std::shared_ptr<PbftVote> &vote, std::string_v
   vote_mgr_->saveOwnVerifiedVote(vote);
 
   LOG(log_nf_) << "Placed " << log_vote_id << " " << vote->getHash() << " for block " << vote->getBlockHash()
-               << ", vote weight " << *vote->getWeight() << ", period " << vote->getPeriod() << ", round "
-               << vote->getRound() << ", step " << vote->getStep();
+               << ", vote weight " << *vote->getWeight() << ", period " << period << ", round " << vote->getRound()
+               << ", step " << vote->getStep();
 
   return true;
 }
@@ -1180,6 +1182,10 @@ PbftManager::proposePbftBlock() {
 
   // Creates pillar block's extra data
   const auto extra_data = createPbftBlockExtraData(current_pbft_period);
+  if (!extra_data.has_value()) {
+    LOG(log_er_) << "Unable to propose pbft block due to wrong pillar block, pbft period: " << current_pbft_period;
+    return {};
+  }
 
   auto ghost = dag_mgr_->getGhostPath(last_period_dag_anchor_block_hash);
   LOG(log_dg_) << "GHOST size " << ghost.size();
@@ -1286,9 +1292,8 @@ std::optional<PbftBlockExtraData> PbftManager::createPbftBlockExtraData(PbftPeri
       return {};
     }
 
-    if (pillar_block->getPeriod() != pbft_period - kGenesisConfig.state.hardforks.ficus_hf.pbft_inclusion_delay) {
+    if (pillar_block->getPeriod() != pbft_period - 1) {
       LOG(log_er_) << "Wrong pillar block period: " << pillar_block->getPeriod() << ", pbft period: " << pbft_period;
-      assert(false);
       return {};
     }
 
@@ -1418,14 +1423,14 @@ bool PbftManager::validatePillarDataInPeriodData(const PeriodData &period_data) 
   const auto block_period = period_data.pbft_blk->getPeriod();
 
   // Validate optional pillar votes presence
-  if (kGenesisConfig.state.hardforks.ficus_hf.isPillarBlockPeriod(block_period, true /* skip first pillar block */)) {
+  if (kGenesisConfig.state.hardforks.ficus_hf.isPbftWithPillarBlockPeriod(block_period)) {
     if (!period_data.pillar_votes_.has_value()) {
-      LOG(log_er_) << "PBFT block " << period_data.pbft_blk->getBlockHash() << ", period " << block_period
+      LOG(log_er_) << "Sync PBFT block " << period_data.pbft_blk->getBlockHash() << ", period " << block_period
                    << " does not contain pillar votes";
       return false;
     }
   } else if (period_data.pillar_votes_.has_value()) {
-    LOG(log_er_) << "PBFT block " << period_data.pbft_blk->getBlockHash() << ", period "
+    LOG(log_er_) << "Sync PBFT block " << period_data.pbft_blk->getBlockHash() << ", period "
                  << period_data.pbft_blk->getPeriod() << " contains pillar votes even though it should not";
     return false;
   }
@@ -1460,47 +1465,22 @@ bool PbftManager::validatePbftBlock(const std::shared_ptr<PbftBlock> &pbft_block
     return false;
   }
 
-  const auto block_period = pbft_block->getPeriod();
-  if (kGenesisConfig.state.hardforks.ficus_hf.isPillarBlockPeriod(block_period, true /* skip first pillar block */)) {
-    const auto last_finalized_pillar_block = pillar_chain_mgr_->getLastFinalizedPillarBlock();
-    if (!last_finalized_pillar_block) {
-      // This should never happen
-      LOG(log_er_) << "Unable to validate PBFT block " << pbft_block_hash << ", period " << block_period
-                   << ". No last finalized pillar block present";
-      return false;
-    }
-
-    if (last_finalized_pillar_block->getPeriod() !=
-        block_period - kGenesisConfig.state.hardforks.ficus_hf.pillar_blocks_interval) {
-      LOG(log_er_) << "Unable to validate PBFT block " << pbft_block_hash << ", period " << block_period
-                   << ". Pillar chain missing finalized block. Last finalized block period "
-                   << last_finalized_pillar_block->getPeriod() << ", expected period "
-                   << block_period - kGenesisConfig.state.hardforks.ficus_hf.pillar_blocks_interval;
-      // Trigger pillar chain votes syncing
-      pillar_chain_mgr_->checkPillarChainSynced(block_period);
-      return false;
-    }
-  }
-
   // Validate optional pillar block hash
+  const auto block_period = pbft_block->getPeriod();
   if (kGenesisConfig.state.hardforks.ficus_hf.isPbftWithPillarBlockPeriod(block_period)) {
     const auto current_pillar_block = pillar_chain_mgr_->getCurrentPillarBlock();
     if (!current_pillar_block) {
       // This should never happen
       LOG(log_er_) << "Unable to validate PBFT block " << pbft_block_hash << ", period " << block_period
                    << ". No current pillar block present in node";
-      assert(false);
       return false;
     }
 
     if (*pbft_block->getExtraData()->getPillarBlockHash() != current_pillar_block->getHash()) {
-      LOG(log_er_) << "PBFT block " << pbft_block_hash << " contains pillar block hash "
-                   << *pbft_block->getExtraData()->getPillarBlockHash()
+      LOG(log_er_) << "PBFT block " << pbft_block_hash << " with period " << pbft_block->getPeriod()
+                   << " contains pillar block hash " << *pbft_block->getExtraData()->getPillarBlockHash()
                    << ", which is different than the local current pillar block" << current_pillar_block->getHash()
-                   << ", period " << pbft_block->getPeriod();
-
-      // Trigger pillar chain votes syncing
-      pillar_chain_mgr_->checkPillarChainSynced(block_period);
+                   << " with period " << current_pillar_block->getPeriod();
       return false;
     }
   }
@@ -1680,10 +1660,33 @@ void PbftManager::finalize_(PeriodData &&period_data, std::vector<h256> &&finali
     }
   }
 
-  const auto result =
+  const auto period = period_data.pbft_blk->getPeriod();
+  auto result =
       final_chain_->finalize(std::move(period_data), std::move(finalized_dag_blk_hashes), std::move(anchor_block));
-  if (synchronous_processing) {
+
+  bool is_pillar_block_period = kGenesisConfig.state.hardforks.ficus_hf.isPillarBlockPeriod(period);
+  if (synchronous_processing || is_pillar_block_period) {
     result.wait();
+
+    if (!is_pillar_block_period) {
+      return;
+    }
+
+    // Create new pillar block
+    const auto finalization_data = result.get();
+    const auto pillar_block = pillar_chain_mgr_->createPillarBlock(finalization_data->final_chain_blk);
+
+    // Check if node is eligible to vote for pillar block
+    // Note: No need to catch ErrFutureBlock because pillar block is create only after pbft block with <period> is
+    // finalized Note: period is used instead of period -1 because at this moment, block with period <period> was
+    // already finalized
+    if (final_chain_->dpos_is_eligible(period, node_addr_)) {
+      if (pillar_block) {
+        pillar_chain_mgr_->genAndPlacePillarVote(period, pillar_block->getHash(), node_sk_, periodDataQueueEmpty());
+      } else {
+        LOG(log_er_) << "Unable to vote for pillar block with period " << period << ". Block was not created";
+      }
+    }
   }
 }
 
@@ -1699,10 +1702,27 @@ bool PbftManager::pushPbftBlock_(PeriodData &&period_data, std::vector<std::shar
     return false;
   }
 
+  auto pbft_period = period_data.pbft_blk->getPeriod();
+
+  // To finalize the pbft block that includes pillar block hash, there needs to be enough pillar votes as well as pbft
+  // votes
+  if (kGenesisConfig.state.hardforks.ficus_hf.isPbftWithPillarBlockPeriod(pbft_period)) {
+    // Note: presence of pillar block hash in extra data was already validated in validatePbftBlock
+    const auto pillar_block_hash = period_data.pbft_blk->getExtraData()->getPillarBlockHash();
+    if (!pillar_chain_mgr_->isPillarBlockLatestFinalized(*pillar_block_hash)) {
+      LOG(log_er_) << "Cannot push PBFT block " << period_data.pbft_blk->getBlockHash() << ", period " << pbft_period
+                   << ". Not enough pillar votes for pillar block " << *pillar_block_hash << ". Request it";
+      if (auto net = network_.lock()) {
+        net->requestPillarBlockVotesBundle(pbft_period - 1, *pillar_block_hash);
+      }
+
+      return false;
+    }
+  }
+
   assert(cert_votes.empty() == false);
   assert(pbft_block_hash == cert_votes[0]->getBlockHash());
 
-  auto pbft_period = period_data.pbft_blk->getPeriod();
   auto null_anchor = period_data.pbft_blk->getPivotDagBlockHash() == kNullBlockHash;
 
   auto batch = db_->createWriteBatch();
@@ -1774,7 +1794,8 @@ PbftPeriod PbftManager::pbftSyncingPeriod() const {
 std::optional<std::pair<PeriodData, std::vector<std::shared_ptr<PbftVote>>>> PbftManager::processPeriodData() {
   auto [period_data, cert_votes, node_id] = sync_queue_.pop();
   auto pbft_block_hash = period_data.pbft_blk->getBlockHash();
-  LOG(log_dg_) << "Pop pbft block " << pbft_block_hash << " from synced queue";
+  LOG(log_dg_) << "Pop pbft block " << pbft_block_hash << " with period " << period_data.pbft_blk->getPeriod()
+               << " from synced queue";
 
   if (pbft_chain_->findPbftBlockInChain(pbft_block_hash)) {
     LOG(log_dg_) << "PBFT block " << pbft_block_hash << " already present in chain.";
@@ -1890,7 +1911,7 @@ std::optional<std::pair<PeriodData, std::vector<std::shared_ptr<PbftVote>>>> Pbf
 
   const auto block_period = period_data.pbft_blk->getPeriod();
   // Validate pillar votes
-  if (kGenesisConfig.state.hardforks.ficus_hf.isPillarBlockPeriod(block_period, true /* skip first pillar block */) &&
+  if (kGenesisConfig.state.hardforks.ficus_hf.isPbftWithPillarBlockPeriod(block_period) &&
       !validatePbftBlockPillarVotes(period_data)) {
     LOG(log_er_) << "Synced PBFT block " << pbft_block_hash << ", period " << block_period
                  << " doesn't have enough valid pillar votes. Clear synced PBFT blocks!";
@@ -1994,8 +2015,7 @@ bool PbftManager::validatePbftBlockPillarVotes(const PeriodData &period_data) co
   }
 
   const auto &pbft_block_hash = period_data.pbft_blk->getBlockHash();
-  const auto required_votes_period =
-      period_data.pbft_blk->getPeriod() - kGenesisConfig.state.hardforks.ficus_hf.pillar_blocks_interval;
+  const auto required_votes_period = period_data.pbft_blk->getPeriod() - 1;
 
   const auto current_pillar_block = pillar_chain_mgr_->getCurrentPillarBlock();
   if (current_pillar_block->getPeriod() != required_votes_period) {
