+++ libraries/core_libs/network/src/tarcap/packets_handlers/latest/get_pbft_sync_packet_handler.cpp
@@ -87,10 +87,10 @@ void GetPbftSyncPacketHandler::sendPbftBlocks(const std::shared_ptr<TaraxaPeer> 
 
     // Add pillar votes to period data
     const auto &ficus_hf_conf = kConf.genesis.state.hardforks.ficus_hf;
-    if (ficus_hf_conf.isPillarBlockPeriod(block_period, true /* skip first pillar block */)) {
+    if (ficus_hf_conf.isPbftWithPillarBlockPeriod(block_period)) {
       // TODO: not ideal solution: should not decode PeriodData, add pillar votes and then encode it again...
       PeriodData period_data{data};
-      auto pillar_data = db_->getPillarBlockData(block_period - ficus_hf_conf.pillar_blocks_interval);
+      auto pillar_data = db_->getPillarBlockData(block_period - 1);
       if (!pillar_data.has_value()) {
         LOG(log_er_) << "DB corrupted. Cannot find pillar votes for period " << block_period << " in db";
         return;
